+++ src/components/request-management/pay-slip/pay-slip-form.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { Button } from '@/components/ui/button';
-import InputField from '@/components/ui/input/input-field';
+import InputField from '@/components/ui/input-field';
 import Spinner from '@/components/ui/spinner';
 import { titleFont } from '@/config/fonts';
 import usePostPaySlip from '@/hooks/request-management/pay-slip/usePostPaySlip';
