+++ src/components/request-management/salary-certificates/salary-certificates-form.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Button } from '@/components/ui/button';
-import InputField from '@/components/ui/input/input-field';
+import InputField from '@/components/ui/input-field';
 import Spinner from '@/components/ui/spinner';
 import { titleFont } from '@/config/fonts';
 import usePostSalaryCetificates from '@/hooks/request-management/salary-certificates/usePostSalaryCetificates';
