+++ src/types/Employee.ts
@@ -1,0 +1,12 @@
+import { Person } from "./Person";
+
+export interface Employee extends Person {
+  id: string;
+  HiringDate: string | Date;
+  NumberChlidren: number;
+  AvailableVacationDays: number;
+  MaritalStatusId: number;
+  GenderId: number;
+  JobPositionId: number;
+  EmbargoId: number;
+}
