+++ src/types/Gender.ts
@@ -1,0 +1,4 @@
+export interface Gender {
+  id: number;
+  Name: string;
+}
