+++ src/types/JobPosition.ts
@@ -1,0 +1,9 @@
+export interface JobPosition {
+  id: number;
+  Name: string;
+  Description: string;
+  baseSalary: string;
+  globalSalary: string;
+  extrafees: string;
+  DepartmentId: number;
+}
