+++ src/types/Login.ts
@@ -1,0 +1,4 @@
+export interface Login {
+  EmployeeId: string;
+  Password: string;
+}
