+++ src/types/MaritalStatus.ts
@@ -1,0 +1,4 @@
+export interface MaritalStatus {
+  id: number;
+  Name: string;
+}
