+++ src/types/Person.ts
@@ -1,0 +1,8 @@
+export interface Person {
+  Name: string;
+  Surname1: string;
+  Surname2: string;
+  Birthdate: string | Date;
+  Email: string;
+  CellPhone: string;
+}
