+++ src/types/UpdateEmployee.ts
@@ -1,0 +1,3 @@
+import { Person } from "./Person";
+
+export interface UpdateEmployee extends Partial<Person> {}
