+++ src/types/User.ts
@@ -1,0 +1,4 @@
+export interface User {
+  EmployeeId: number;
+  Password: string;
+}
