+++ src/types/auth/ChangePassword.ts
@@ -1,0 +1,5 @@
+export interface ChangePassword {
+  OldPassword: string;
+  Password: string;
+  ConfirmPassword: string;
+}
