+++ src/types/auth/ForgotPassword.ts
@@ -1,0 +1,3 @@
+export interface ForgotPassword {
+  Email: string;
+}
