+++ src/types/auth/ResetPassword.ts
@@ -1,0 +1,3 @@
+export interface ResetPassword {
+  Password: string;
+}
