+++ src/types/index.ts
@@ -1,0 +1,15 @@
+export * from './Person';
+export * from './Employee';
+export * from './UpdateEmployee';
+export * from './Gender';
+export * from './MaritalStatus';
+export * from './User';
+export * from './Login';
+
+export * from './auth/ForgotPassword';
+export * from './auth/ResetPassword';
+export * from './auth/ChangePassword';
+
+export * from './request/RequestSalaryCertificate';
+export * from './request/RequestPaySlip';
+export * from './request/RequestVacation';
