+++ src/types/request/RequestPaySlip.ts
@@ -1,0 +1,5 @@
+export interface RequestPaySlip {
+  date: Date;
+  reason: string;
+  employeeId: number;
+}
