+++ src/types/request/RequestSalaryCertificate.ts
@@ -1,0 +1,5 @@
+export interface RequestSalaryCertificate {
+  employeeId: number;
+  requestDate: Date;
+  reason: string;
+}
