+++ src/types/request/RequestVacation.ts
@@ -1,0 +1,9 @@
+export interface RequestVacation {
+  daysRequested: number;
+  departureDate: Date;
+  entryDate: Date;
+  departmentApproval: boolean;
+  RRHHApproval: boolean;
+  mayorApproval: boolean;
+  RequestId: number;
+}
